+++ src/repositories/reportCard/index.ts
@@ -1,0 +1,1 @@
+export * from "@/repositories/reportCard/reportCard.repository";